+++ Sources/Extensions/DateFormattingService/DateFormattingService+DefaultImplementation.swift
@@ -25,26 +25,31 @@ import SwiftDate
 public extension DateFormattingService {
 
     func date(from string: String,
-              format: String,
+              format: DateFormatType,
               defaultDate: DateInRegion = Date().inDefaultRegion()) -> DateInRegion {
 
-        return date(from: string, format: format) ?? defaultDate
+        return date(from: string, format: format, parsedIn: nil) ?? defaultDate
     }
 
-    func date(from string: String, format: String) -> DateInRegion? {
-        return DateInRegion(string, format: format, region: currentRegion)
+    func date(from string: String,
+              format: DateFormatType,
+              parsedIn: Region?) -> DateInRegion? {
+
+        let region = parsedIn ?? currentRegion
+
+        return format.stringToDateFormat.toDate(string, region: region)
     }
 
-    func string(from date: DateInRegion, format: DateFormatType) -> String {
-        return date.toString(format.swiftDateFormat)
+    func string(from date: DateRepresentable, format: DateFormatType) -> String {
+        return format.dateToStringFormat.toString(date)
     }
 
-    func string(from date: DateInRegion, format: DateFormatType, formattedIn: Region? = nil) -> String {
+    func string(from date: DateRepresentable, format: DateFormatType, formattedIn: Region?) -> String {
         let region = formattedIn ?? currentRegion
 
         let dateInFormatterRegion = date.convertTo(region: region)
 
-        return dateInFormatterRegion.toString(format.swiftDateFormat)
+        return format.dateToStringFormat.toString(dateInFormatterRegion)
     }
 
 }
@@ -60,7 +65,7 @@ public extension DateFormattingService where Self: Singleton {
     ///   - defaultDate: Default date if formatting will fail.
     /// - Returns: Date parsed from given string or default date if parsing did fail.
     static func date(from string: String,
-                     format: String,
+                     format: DateFormatType,
                      defaultDate: DateInRegion = Date().inDefaultRegion()) -> DateInRegion {
 
         return shared.date(from: string, format: format, defaultDate: defaultDate)
@@ -71,9 +76,10 @@ public extension DateFormattingService where Self: Singleton {
     /// - Parameters:
     ///   - string: String to use for date parsing.
     ///   - format: Format that should be used for date parsing.
+    ///   - parsedIn: A region that should be used for date parsing. In case of nil defaultRegion will be used.
     /// - Returns: Date parsed from given string or default date if parsing did fail.
-    static func date(from string: String, format: String) -> DateInRegion? {
-        return shared.date(from: string, format: format)
+    static func date(from string: String, format: DateFormatType, parsedIn: Region?) -> DateInRegion? {
+        return shared.date(from: string, format: format, parsedIn: parsedIn)
     }
 
     /// Method format date in given format.
@@ -82,7 +88,7 @@ public extension DateFormattingService where Self: Singleton {
     ///   - date: Date to format.
     ///   - format: Format that should be used for date formatting.
     /// - Returns: String that contains formatted date or nil if formatting did fail.
-    static func string(from date: DateInRegion, format: DateFormatType) -> String {
+    static func string(from date: DateRepresentable, format: DateFormatType) -> String {
         return shared.string(from: date, format: format)
     }
 
@@ -93,7 +99,7 @@ public extension DateFormattingService where Self: Singleton {
     ///   - format: Format that should be used for date formatting.
     ///   - formattedIn: A region that should be used for date formatting. In case of nil defaultRegion will be used.
     /// - Returns: String that contains formatted date or nil if formatting did fail.
-    static func string(from date: DateInRegion, format: DateFormatType, formattedIn: Region?) -> String {
+    static func string(from date: DateRepresentable, format: DateFormatType, formattedIn: Region?) -> String {
         return shared.string(from: date, format: format, formattedIn: formattedIn)
     }